bot_app/requirements.txt
botbuilder-core
aiohttp
openai
requests
python-dotenv
apscheduler
